Validate orders and order tools
Using the `order_creator.rb` PORO, orders and order tools are validated before
being saved.

Specifically, an order cannot be saved without being associated with a user. An
order_tool cannot be saved without being associated with an order.

As a result, it is not possible to created orphaned orders or order tools.