Remove cart_login method from SessionsController
Previously we had created a login view specifically for users who logged in
from the cart view.

This commit refactors the SessionsController so that now all users use the
standard `new.html.erb` view through the `new` method by using a new instance
variable that is true if the referring path is "/cart/login", thus triggering
a hidden field in the view.