Send admin to admin_dashboard_path
This commit changes the `show` and `update` methods in the UsersController. It
inserts conditionals into both of these methods so that if the requesting
user is an admin, the user is taken to the admin_dashboard_path when redirected
instead of to the user dashboard view.

Getting all the tests to pass meant changing an assertion in a previous test so
that this improved flow would be accounted for in that test.