Allow admin to suspend and reactivate a home
This commit creates a new column in the homes table that indicates whether a
home is activate or suspend. This state is controlled and accessed by new
methods in the homes model.

An admin can control this state through the admin owners page, and a user can
either reserve or a home or is told that the home is unavailable based on this
state.